+++ src/main/java/com/mraof/minestuck/item/GutterBallItem.java
@@ -1,6 +1,7 @@
 package com.mraof.minestuck.item;
 
 import com.mraof.minestuck.player.PlayerSavedData;
+import net.minecraft.ChatFormatting;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
@@ -23,6 +24,7 @@ import java.util.List;
 @MethodsReturnNonnullByDefault
 public class GutterBallItem extends Item
 {
+	public static final String MINOR_INCREASE = "message.gutter.minor_increase";
 	public GutterBallItem(Properties properties)
 	{
 		super(properties);
@@ -34,6 +36,7 @@ public class GutterBallItem extends Item
 		ItemStack itemStack = player.getItemInHand(usedHand);
 		level.playSound(player, player.getX(), player.getY(), player.getZ(), SoundEvents.BELL_RESONATE, SoundSource.PLAYERS, 0.5F, 0.3F);
 		itemStack.shrink(1);
+		player.displayClientMessage(Component.translatable(MINOR_INCREASE).withStyle(ChatFormatting.BOLD), true);
 		
 		if(player instanceof ServerPlayer serverPlayer)
 		{