+++ src/main/java/com/mraof/minestuck/item/loot/GristLootModifier.java
@@ -1,10 +1,10 @@
 package com.mraof.minestuck.item.loot;
 
-import com.google.common.collect.Lists;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import com.mraof.minestuck.alchemy.GristCostRecipe;
 import com.mraof.minestuck.alchemy.GristSet;
+import com.mraof.minestuck.alchemy.recipe.GristCostRecipe;
+import com.mraof.minestuck.entity.item.GristEntity;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.LootContext;
@@ -49,10 +49,10 @@ public class GristLootModifier extends LootModifier
 			{
 				GristSet cost = GristCostRecipe.findCostForItem(stack, null, true, context.getLevel());
 				if(cost != null && multiplier != 1)
-					cost = new GristSet(cost).scale(multiplier, false);
+					cost = cost.mutableCopy().scale(multiplier, false);
 				
 				if(cost != null && !cost.isEmpty())
-					cost.spawnGristEntities(context.getLevel(), pos.x, pos.y, pos.z,
+					GristEntity.spawnGristEntities(cost, context.getLevel(), pos.x, pos.y, pos.z,
 							context.getRandom(), entity -> entity.setDeltaMovement(entity.getDeltaMovement().multiply(0.5, 0.5, 0.5)));
 				else
 					remainingLoot.add(stack);