+++ src/main/java/com/mraof/minestuck/util/MSSoundEvents.java
@@ -100,6 +100,9 @@ public class MSSoundEvents
 	public static final RegistryObject<SoundEvent> EVENT_EDIT_TOOL_REVISE = register("event.edit_tools.revise");
 	public static final RegistryObject<SoundEvent> EVENT_EDIT_TOOL_RECYCLE = register("event.edit_tools.recycle");
 	public static final RegistryObject<SoundEvent> CRUXTRUDER_DOWEL = register("block.cruxtruder.dowel");
+	public static final RegistryObject<SoundEvent> ALCHEMITER_RESONATE = register("block.alchemiter.resonate");
+	public static final RegistryObject<SoundEvent> TRANSPORTALIZER_TELEPORT = register("block.transportalizer.teleport");
+	public static final RegistryObject<SoundEvent> TOTEM_LATHE_LATHE = register("block.totem_lathe.lathe");
 	public static final RegistryObject<SoundEvent> BLOCK_HORSE_CLOCK_CHIME = register("block.horse_clock.chime");
 	public static final RegistryObject<SoundEvent> BLOCK_CLOCK_TICK = register("block.clock.tick");
 	public static final RegistryObject<SoundEvent> BLOCK_CLOCK_TOCK = register("block.clock.tock");